requirements.txt
asgiref==3.8.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-config-url==0.1.1
Django==5.1.6
django-allauth==65.13.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
idna==3.11
inflection==0.5.1
jwcrypto==1.5.6
oauthlib==3.3.1
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
random-username==1.0.2
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.1.1
urllib3==2.5.0
